# Conflicts:
#	.idea/workspace.xml

+++ Code/Controller/ControllerParametre_Tuteur.php
@@ -32,7 +32,7 @@ if (unserialize($_SESSION["utilisateur"]) instanceof Administrateur) {
         }else {
             $Message = "";
         }
-        if (isset($_POST['preTut']) && isset($_POST['nomTut']) && isset ($_POST['telTut']) && isset($_POST['adrTut']) && isset($_POST['vilTut']) && isset($_POST['cpTut']) && isset($_POST['mailTut']) && isset($_POST['logTut']) && isset($_POST['mdpTut']) && empty($_POST['tuteur-select'])) {
+        if ($_POST['preTut'] != "" && $_POST['nomTut'] != "" && $_POST['telTut'] != "" && $_POST['adrTut'] != "" && $_POST['vilTut'] != "" && $_POST['cpTut'] != "" && $_POST['mailTut'] != "" && $_POST['logTut'] != "" && $_POST['mdpTut'] != "" && empty($_POST['tuteur-select'])) {
             $pre = $_POST['preTut'];
             $nom = $_POST['nomTut'];
             $tel = $_POST['telTut'];
@@ -46,9 +46,13 @@ if (unserialize($_SESSION["utilisateur"]) instanceof Administrateur) {
             $nbMax4 = ($_POST['nbMaxEtu4'] > 0) ? $_POST['nbMaxEtu4'] : 0;
             $nbMax5 = ($_POST['nbMaxEtu5'] > 0) ? $_POST['nbMaxEtu5'] : 0;
             $tuteur = new Tuteur($nbMax3, $nbMax4, $nbMax5, 0, $log, $mdp, $mail, $tel, $nom, $pre, $adr, $cp, $vil);
-            if($tutDAO->create($tuteur)){
-                $Message = "";
-                $tuts = $tutDAO->getAll();
+            if($tutDAO->verifLog($log)){
+                $Message = "Log deja utiliser";
+            } else {
+                if($tutDAO->create($tuteur)){
+                    $Message = "le tuteur a bien été créé";
+                    $tuts = $tutDAO->getAll();
+                }
             }
         } else {
             $Message = "veuillez remplir tous les champs";
@@ -56,34 +60,44 @@ if (unserialize($_SESSION["utilisateur"]) instanceof Administrateur) {
     }
     //---------------------- modifier ---------------------------
     if (isset($_POST['btnUpdate'])){
-        if (isset($_POST['preTut']) && isset($_POST['nomTut']) && isset ($_POST['telTut']) && isset($_POST['adrTut']) && isset($_POST['vilTut']) && isset($_POST['cpTut']) && isset($_POST['mailTut']) && isset($_POST['logTut']) && isset($_POST['mdpTut']) && isset($_POST['tuteur-select'])){
-            $tuteur = $tutDAO->find($_POST["tuteur-select"]);
-            $pre = $_POST['preTut'];
-            $nom = $_POST['nomTut'];
-            $tel = $_POST['telTut'];
-            $adr = $_POST['adrTut'];
-            $vil = $_POST['vilTut'];
-            $cp = $_POST['cpTut'];
-            $mail = $_POST['mailTut'];
-            $log = $_POST['logTut'];
-            $mdp = $_POST['mdpTut'];
-            $nbMax3 = ($_POST['nbMaxEtu3'] > 0) ? $_POST['nbMaxEtu3'] : 0;
-            $nbMax4 = ($_POST['nbMaxEtu4'] > 0) ? $_POST['nbMaxEtu4'] : 0;
-            $nbMax5 = ($_POST['nbMaxEtu5'] > 0) ? $_POST['nbMaxEtu5'] : 0;
-            $tuteur->setPreUti($pre);
-            $tuteur->setNomUti($nom);
-            $tuteur->setTelUti($tel);
-            $tuteur->setAdrUti($adr);
-            $tuteur->setVilUti($vil);
-            $tuteur->setCpUti($cp);
-            $tuteur->setMailUti($mail);
-            $tuteur->setLogUti($log);
-            $tuteur->setMdpUti($mdp);
-            $tuteur->setNbMax3($nbMax3);
-            $tuteur->setNbMax4($nbMax4);
-            $tuteur->setNbMax5($nbMax5);
-            if($tutDAO->update($tuteur)){
-                $Message = "Le tuteur a bien été modifier";
+        if ($_POST['preTut'] != "" && $_POST['nomTut'] != "" && $_POST['telTut'] != "" && $_POST['adrTut'] != "" && $_POST['vilTut'] != "" && $_POST['cpTut'] != "" && $_POST['mailTut'] != "" && $_POST['logTut'] != "" && $_POST['mdpTut'] != "" && isset($_POST['tuteur-select'])){
+            if ($_POST['tuteur-select'] != ''){
+                $tuteur = $tutDAO->find($_POST["tuteur-select"]);
+                $pre = $_POST['preTut'];
+                $nom = $_POST['nomTut'];
+                $tel = $_POST['telTut'];
+                $adr = $_POST['adrTut'];
+                $vil = $_POST['vilTut'];
+                $cp = $_POST['cpTut'];
+                $mail = $_POST['mailTut'];
+                $log = $_POST['logTut'];
+                $mdp = $_POST['mdpTut'];
+                $nbMax3 = ($_POST['nbMaxEtu3'] > 0) ? $_POST['nbMaxEtu3'] : 0;
+                $nbMax4 = ($_POST['nbMaxEtu4'] > 0) ? $_POST['nbMaxEtu4'] : 0;
+                $nbMax5 = ($_POST['nbMaxEtu5'] > 0) ? $_POST['nbMaxEtu5'] : 0;
+                if ($tuteur->getLogUti() != $log){                                      //verif si ca a changer : true
+                    if ($tutDAO->verifLog($log)){
+                        $Message = "Log deja utiliser";
+                    } else {
+                        $tuteur->setPreUti($pre);
+                        $tuteur->setNomUti($nom);
+                        $tuteur->setTelUti($tel);
+                        $tuteur->setAdrUti($adr);
+                        $tuteur->setVilUti($vil);
+                        $tuteur->setCpUti($cp);
+                        $tuteur->setMailUti($mail);
+                        $tuteur->setLogUti($log);
+                        $tuteur->setMdpUti($mdp);
+                        $tuteur->setNbMax3($nbMax3);
+                        $tuteur->setNbMax4($nbMax4);
+                        $tuteur->setNbMax5($nbMax5);
+                        if($tutDAO->update($tuteur)){
+                            $Message = "Le tuteur a bien été modifier";
+                        }
+                    }
+                }
+            } else {
+                $Message = "Ce tuteur n'existe pas";
             }
         } else {
             $Message = "Veuillez selectionner un tuteur à modifier";
@@ -94,18 +108,25 @@ if (unserialize($_SESSION["utilisateur"]) instanceof Administrateur) {
         if (isset($_POST['tuteur-select'])){
             if ($_POST['tuteur-select'] != ''){
                 $tuteur = $tutDAO->find($_POST["tuteur-select"]);
-                if ($tutDAO->delete($tuteur)){
-                    $Message = "suppression ok";
-                    $tuts = $tutDAO->getAll();
+                if ($tuteur){
+                    if ($tutDAO->delete($tuteur)){
+                        $Message = 'suppression du tuteur ' . $tuteur->getNomUti();
+                        $tuts = $tutDAO->getAll();
+                    }
+                } else {
+                    $Message = "Ce tuteur n'existe pas";
                 }
+
             } else {
-                $Message = "veuillez selectionner un tuteur";
+                $Message = "veuillez selectionner un tuteur a supprimer";
             }
         } else {
-            $Message = "veuillez selectionner un tuteur";
+            $Message = "veuillez selectionner un tuteur a supprimer";
         }
     }
 
     include_once ('../View/header_admin.php');
     include_once ('../View/Page_Parametre_Tuteur.php');
+} else {
+    header('Location: ControllerConnexion.php');
 }
